1480. Running Sum of 1d Array

+++ c++/1480-running-sum-of-1d-array.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
-    // O(n) time | O(1) space - n is the length of nums
+    // O(n) time | O(1) space
     vector<int> runningSum(vector<int>& nums) {
-        int n = nums.size();
-        for (int i = 1; i < n; ++i) {
+        size_t n = nums.size();
+        for (size_t i = 1; i < n; ++i) {
             nums[i] += nums[i - 1];
         }
         return nums;
